// src/api/generated/schemas/registerEmpRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface RegisterEmpRequest {
  /**
   * 성명
   * @minLength 0
   * @maxLength 50
   */
  empNm: string;
  /**
   * 사내 계정 ID (UK)
   * @minLength 0
   * @maxLength 50
   */
  worldId?: string;
  /**
   * 이메일
   * @minLength 0
   * @maxLength 100
   */
  emailAddr?: string;
  /**
   * 연락처
   * @minLength 0
   * @maxLength 30
   */
  telNo?: string;
  /** DEPT ID */
  deptId?: number;
}